[SPARK-GREENPLUM-4]Add transactions support for transmitting data to greenplum.
---

What changes were proposed in this pull request?
Add transactions support for transmitting data from sparkSql to greenplum.
There are several cases of data transmission:

case1: Append data to origin gptable.
case2: Overwrite origin gptable, but the table is a cascadingTruncateTable, so we can not drop the gptable, we have to truncate it and append data.
case3: Overwrite origin table and the table is not a cascadingTruncateTable, so we can drop it.
I add a transactionOn option in GreenplumOptions, default is false.
The copy operation to greenplum for each partition is atomic.
For case1 and case2, when transactionOn is true, we will coalesce the dataframe's partitions to one partition, then the operation of the only partition is an atomic operation.
Otherwise, the copy operation for all partitions is not in a single transaction.

For case3, we create a temp greenplum table, and copy each partition's data to the temp table.
We register an accmulator to record the number of successful operation, which copies a partition's data to temp table.
If all operations for each partition are successful, we drop the gptable if it exists, and rename the temp table to gptable.
Otherwise, there are some unsuccessful partitions, we will drop the temp table and throw an exception to notify the user.

How was this patch tested?
Unit test.

---

fix #7 fix#8

---

Squashed commit of the following:

commit 02d60b83b926437ebb124711931865c8c5a7f1a3
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 16:28:08 2019 +0800

    add max retry times for drop the temp table

commit 6c6941f8761f3651677ad6f420e3ebcdf43d5fb7
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 15:55:23 2019 +0800

    catch exception when drop the temp table

commit 14999b7bb85f940582c4b6aa54aa6c74472b0ffe
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 15:22:54 2019 +0800

     fix code

commit d2ad86722fd167ba6874c3408a0e137a459fa0ec
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 15:14:46 2019 +0800

    modify the notification message

commit 539c49cbbd8891d145499cac65fda2e73b553bde
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 14:06:19 2019 +0800

    classsify copyToGp operation to transactionalCopy and nonTransactionalCopy

commit bb02867a6459fd2cc49be6a2293dac2d8b4cc420
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 11:53:18 2019 +0800

    some function optimization

commit 5bc3bd573ff8a3ea2aa26997dceca34e844be426
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 11:27:22 2019 +0800

    fix code

commit 9fb564e951ef0f95d69dafdf32e21a4b26e20731
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 11:18:21 2019 +0800

    fix unit test

commit 2e45d25d251919fe9b2dcc9a47f21a2f21b58e5a
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 11:01:35 2019 +0800

    define partitionCopyFailuerException and throws exception when some partitions failed

commit b9086aabb03c25cda15c7c49e978a9cf5b79f61c
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 10:30:11 2019 +0800

    fix code and add transactionForAppend option

commit 250e0ac5d004cde8e8c5a9d0f71fe1d45d7f023c
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Thu May 9 20:47:24 2019 +0800

    modify unit test

commit 5fd569a19d90f59dfba770f7a1508202e1377fe1
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Thu May 9 18:33:50 2019 +0800

    fix code 2

commit 6873757137f4794cc43027a4774dd32512f855e9
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Thu May 9 16:11:31 2019 +0800

    fix code

commit ec0a349fbcc8257676568c590440f33b30923b35
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Wed May 8 15:38:47 2019 +0800

    [SPARK-GREENPLUM-4]Add transactions support for transmitting data to greenplum.