[SPARK-GREENPLUM-5] Set a timeout for copy operation, which copies a partition's data to greenplum.
---

What changes were proposed in this pull request?
I add a copyTimeout option to GreenplumOptions.
And for each partition, I start a copy thread and check the copy thread's state.
The exception of copy thread would be throwed and if the copy thread has been running for more than the copyTimeout, a timeout exception would be throwed.
Otherwise, the copy operation is completed in timeout.

How was this patch tested?
Existing unit test.

---
 fix #9 fix #10
---

Squashed commit of the following:

commit e6a39a7b9981a2577695c78bc5daf4e469d0bdb1
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Mon May 13 14:54:22 2019 +0800

     turn down the check interval

commit bc2b73815f7f9753d69cabf2dce6ebd4873d9c3b
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Mon May 13 14:21:37 2019 +0800

    fix code

commit 9bd6c32cca01db6e1ae243d06dc89a6d27b08c3b
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Mon May 13 14:06:23 2019 +0800

    use timeStringAsMs

commit 13bfda13687458e85130ed9def639f5d8ec08451
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Mon May 13 12:33:52 2019 +0800

    remove callable

commit 6e7ad15f9a381de3b3a2dd58c1e5cd5903584af2
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 10 19:26:41 2019 +0800

    [SPARK-GREENPLUM-5] Set a timeout of cppy operation for copying a partition's data to greenplum.