[SPARK-GREENPLUM-9] Fix the bug when creating temp table.
---
What changes were proposed in this pull request?
In a GreenplumOptions, if its table option is in a format that both schema and rawTableName are closured by double quotes.
Such as :

"schemaName"."tableName"
So we can not append a randomString on the end of options.table directly.

How was this patch tested?
Unit test.
---
fix #14
---
Squashed commit of the following:

commit 2c4ff8800c1b73b24249cc82dada67b3acbd0af5
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Sat May 25 00:03:52 2019 +0800

    fix nit

commit 53dc84eb34131398b1f7047595b91addc32ba71a
Author: hustfeiwang <wangfei3@corp.netease.com>
Date:   Fri May 24 17:00:02 2019 +0800

    [SPARK-GREENPLUM-9]Fix the bug when creating temp table.